+++ src/components/ac-polyline/ac-polyline.component.ts
@@ -5,9 +5,8 @@ import { PolylineDrawerService } from '../../services/drawers/polyline-drawer/po
 /**
  *  This is a polyline implementation.
  *  The ac-label element must be a child of ac-map element.
- *  The properties of props are the same as the properties of Entity and PolylineGraphics:
- *  + https://cesiumjs.org/Cesium/Build/Documentation/Entity.html
- *  + https://cesiumjs.org/Cesium/Build/Documentation/PolylineGraphics.html
+ *  The properties of props are the same as the properties of Polyline Primitive:
+ *  + https://cesiumjs.org/Cesium/Build/Documentation/Polyline.html
  *
  *  __Usage:__
  *  ```
@@ -15,6 +14,7 @@ import { PolylineDrawerService } from '../../services/drawers/polyline-drawer/po
  *    position: position,
  *    text: 'labelText',
  *    font: '30px sans-serif'
+ *    color: Cesium.Color.GREEN
  *  }">;
  *  </ac-polyline>
  *  ```